+++ mayan/apps/templating/tests/test_tags.py
@@ -1,6 +1,18 @@
+from datetime import datetime
+
 from mayan.apps.testing.tests.base import BaseTestCase
 
 from .mixins import TemplateTagTestMixin
+
+
+class TemplateTagDateParseTestCase(TemplateTagTestMixin, BaseTestCase):
+    def test_basic_functionality(self):
+        now = datetime.now()
+
+        result = self._render_test_template(
+            template_string='{{% set "{}"|date_parse as date_object %}}{{{{ date_object.year }}}}'.format(now.isoformat())
+        )
+        self.assertEqual(result, str(now.year))
 
 
 class TemplateFilterDictGetTestCase(TemplateTagTestMixin, BaseTestCase):
@@ -99,3 +111,13 @@ class TemplateTagSetTestCase(TemplateTagTestMixin, BaseTestCase):
             template_string='{% set nonexistant as result %}{{ result }}'
         )
         self.assertEqual(result, '')
+
+
+class TemplateTagTimeDeltaTestCase(TemplateTagTestMixin, BaseTestCase):
+    def test_basic_functionality(self):
+        now = datetime.now()
+
+        result = self._render_test_template(
+            template_string='{{% set "{}"|date_parse as date_object %}}{{% timedelta date_object days=366 as date_new %}}{{{{ date_new.year }}}}'.format(now.isoformat())
+        )
+        self.assertEqual(result, str(now.year + 1))